requirements.txt
flask==3.1.3
python-dotenv>=1.0.0
google-genai>=1.0.0
gunicorn>=22.0.0
pymupdf>=1.24.0
requests>=2.32.0
